experiment-lib: Create separate base class for experiments
This will allow us to more flexibly create multiple different kinds of
experiments for different benchmarks.
