+++ experiments/__init__.py
@@ -1,6 +1,7 @@
 """Specify and run experiments in the simulator and Spark service."""
 
 from .experiment_spec import Experiment
+from .tpch_experiment import TpchExperiment
 from .scheduler_spec import SchedSpec, sched_specs
 from .simulator_experiments import run_simulator_experiment
 from .service_experiments import run_service_experiment
